Use owner, repo, issue_number instead of context
octokit throws a warning when `number` is specified instead of `issue_number`. The `context.issues` method is a helper that hardcodes the `number` field in its return value:

https://github.com/actions/toolkit/blob/4e69ce10e91ecd239e16343fe2ae7e54d3e35fed/packages/github/src/context.ts#L42-L49

`issue()` and `repo()` are the only getter methods serving as helpers. I don't know if there are any plans to expand this in the future, but knowing how to calculate these from scratch is useful in the case they go away!